Quick fix for the motion issue.
This bug seems to come from baxter update. Now the joint trajectory
controller based on actionlib requires the first point of a goal trajectory
that has multiple waypoints to be very close to current position.

The fix (also implemented in [1]) is to set the current position of the
robot as the first waypoint.

[1](https://github.com/RethinkRobotics/baxter_examples/blob/master/scripts/joint_trajectory_client.py#L139-L140)